requirements.txt
django
mysqlclient
djangorestframework
pygments
djangorestframework-jsonapi
djangorestframework